delete: Motorcycle, Vehicle, Person, Employee classes
