--- CarRental/CarRental.DAL/Models/Employee.cs
@@ -1,5 +1,0 @@
-﻿namespace CarRental.DAL.Models;
-public sealed class Employee
-{
-    public int EmployeeId { get; set; }
-}
