--- CarRental/CarRental.DAL/Models/Motorcycle.cs
@@ -1,10 +1,0 @@
-﻿namespace CarRental.DAL.Models;
-public sealed class Motorcycle : Vehicle
-{
-    public int WeightInKg;
-
-    public override string ToString()
-    {
-        return base.ToString();
-    }
-}
